Change AUTO_OFFSET_RESET_CONFIG of OppfolgingstilfelleConsumer to latest
If there is  no initial offset in Kafka or if the current offset does not exist any more on the server, then the consumer will begin at the latest offset, not the earliest. This is to prevent consuming multiple times in case of failure and potential skipped message are torable for isprediksjon.
Also enable manual commit for topic Oppfolgingstilfelle.